インスタンスを持たないクラスを削除
