+++ 08.ls_object/ls_command_oop.rb
@@ -5,10 +5,5 @@ require 'pathname'
 require 'etc'
 require 'date'
 require './file'
-require './permission'
-require './list'
-require './option_command'
-require './regular_expression'
-require './calculation'
 
 Filedata.output
